Moved database to root, merged various utils into helpers.py
